move database files
